renamed file candidate_cell_mapping to coarse_lookup_map to match class name
