+++ CMakeLists.txt
@@ -125,7 +125,7 @@ set(HEADER_FILES
     ${INC_DIR}/nektar_interface/geometry_transport_3d.hpp
     ${INC_DIR}/nektar_interface/halo_extension.hpp
     ${INC_DIR}/nektar_interface/particle_boundary_conditions.hpp
-    ${INC_DIR}/nektar_interface/particle_cell_mapping/candidate_cell_mapping.hpp
+    ${INC_DIR}/nektar_interface/particle_cell_mapping/coarse_lookup_map.hpp
     ${INC_DIR}/nektar_interface/particle_cell_mapping/generated_linear/hexahedron.hpp
     ${INC_DIR}/nektar_interface/particle_cell_mapping/generated_linear/linear_newton_implementation.hpp
     ${INC_DIR}/nektar_interface/particle_cell_mapping/generated_linear/prism.hpp
