+++ include/nektar_interface/particle_cell_mapping/particle_cell_mapping.hpp
@@ -9,7 +9,7 @@
 
 #include "../coordinate_mapping.hpp"
 #include "../particle_mesh_interface.hpp"
-#include "candidate_cell_mapping.hpp"
+#include "coarse_lookup_map.hpp"
 #include "particle_cell_mapping_2d.hpp"
 #include "particle_cell_mapping_3d.hpp"
