+++ include/nektar_interface/particle_cell_mapping/particle_cell_mapping_2d.hpp
@@ -10,7 +10,7 @@
 #include "../coordinate_mapping.hpp"
 #include "../geometry_transport_2d.hpp"
 #include "../particle_mesh_interface.hpp"
-#include "candidate_cell_mapping.hpp"
+#include "coarse_lookup_map.hpp"
 #include "newton_geom_interfaces.hpp"
 #include "particle_cell_mapping_common.hpp"
